+++ build_runner/lib/src/entrypoint/options.dart
@@ -2,18 +2,25 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 import 'dart:async';
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
-import 'package:build_runner/build_runner.dart';
+import 'package:glob/glob.dart';
+import 'package:io/io.dart';
 import 'package:meta/meta.dart';
+import 'package:path/path.dart' as p;
 import 'package:shelf/shelf_io.dart';
+
+import 'package:build_runner/build_runner.dart';
 
 const _assumeTty = 'assume-tty';
 const _deleteFilesByDefault = 'delete-conflicting-outputs';
 const _failOnSevere = 'fail-on-severe';
 const _hostname = 'hostname';
+
+final _pubBinary = Platform.isWindows ? 'pub.bat' : 'pub';
 
 /// Unified command runner for all build_runner commands.
 class BuildCommandRunner extends CommandRunner {
@@ -25,6 +32,7 @@ class BuildCommandRunner extends CommandRunner {
     addCommand(new _BuildCommand());
     addCommand(new _WatchCommand());
     addCommand(new _ServeCommand());
+    addCommand(new _TestCommand());
   }
 }
 
@@ -109,12 +117,15 @@ abstract class _BaseCommand extends Command {
       (runner as BuildCommandRunner).builderApplications;
 
   _BaseCommand() {
+    _addBaseFlags();
+  }
+
+  void _addBaseFlags() {
     argParser
       ..addFlag(_assumeTty,
           help: 'Enables colors and interactive input when the script does not'
               ' appear to be running directly in a terminal, for instance when it'
               ' is a subprocess',
-          defaultsTo: null,
           negatable: true)
       ..addFlag(_deleteFilesByDefault,
           help:
@@ -213,5 +224,108 @@ class _ServeCommand extends _WatchCommand {
     }
     await handler.buildResults.drain();
     await Future.wait(servers.map((server) => server.close()));
+  }
+}
+
+/// A [Command] that does a single build and then runs tests using the compiled
+/// assets.
+class _TestCommand extends _BaseCommand {
+  @override
+  final argParser = new ArgParser(allowTrailingOptions: false);
+
+  @override
+  String get name => 'test';
+
+  @override
+  String get description =>
+      'Performs a single build on the specified targets and then runs tests '
+      'using the compiled assets.';
+
+  @override
+  Future<Null> run() async {
+    var options = _readOptions();
+    await build(builderApplications,
+        deleteFilesByDefault: options.deleteFilesByDefault,
+        assumeTty: options.assumeTty);
+
+    // Create the merged output directory.
+    String precompiledPath = await _createMergedDir();
+
+    // Create missing *_test.html files.
+    await _createMissingHtmlFiles(precompiledPath);
+
+    // Run the tests!
+    await _runTests(precompiledPath);
+
+    await ProcessManager.terminateStdIn();
+  }
+
+  /// Creates a merged directory for the current build script.
+  Future<String> _createMergedDir() async {
+    var tmpDir = await Directory.systemTemp.createTemp('build_runner_test');
+    var tmpDirPath = tmpDir.absolute.uri.toFilePath();
+    var scriptLocation = Platform.script.toFilePath();
+    var process = await new ProcessManager().spawn(_pubBinary, [
+      'run',
+      'build_runner:create_merged_dir',
+      '--script',
+      scriptLocation,
+      '--output-dir',
+      tmpDirPath,
+    ]);
+
+    var processExitCode = await process.exitCode;
+    if (processExitCode != 0) {
+      print('Error creating merged dir! :(');
+      exit(processExitCode);
+    }
+    return tmpDirPath;
+  }
+
+  /// Creates html files for tests in [tmpDirPath] that are missing them.
+  Future<Null> _createMissingHtmlFiles(String tmpDirPath) async {
+    var dartBrowserTestSuffix = '_test.dart.browser_test.dart';
+    var htmlTestSuffix = '_test.html';
+    var dartFiles =
+        new Glob('test/**$dartBrowserTestSuffix').list(root: tmpDirPath);
+    await for (var file in dartFiles) {
+      var dartPath = p.relative(file.path, from: tmpDirPath);
+      var htmlPath = dartPath.substring(
+              0, dartPath.length - dartBrowserTestSuffix.length) +
+          htmlTestSuffix;
+      var htmlFile = new File(p.join(tmpDirPath, htmlPath));
+      if (!await htmlFile.exists()) {
+        var originalDartPath = p.basename(dartPath.substring(
+            0, dartPath.length - '.browser_test.dart'.length));
+        await htmlFile.writeAsString('''
+<!DOCTYPE html>
+<html>
+<head>
+  <title>${HTML_ESCAPE.convert(htmlPath)} Test</title>
+  <link rel="x-dart-test"
+        href="${HTML_ESCAPE.convert(originalDartPath)}">
+  <script src="packages/test/dart.js"></script>
+</head>
+</html>''');
+      }
+    }
+  }
+
+  /// Runs tests using [precompiledPath] as the precompiled test directory.
+  Future<Null> _runTests(String precompiledPath) async {
+    var extraTestArgs = argResults.rest;
+    var testProcess = await new ProcessManager().spawn(
+        _pubBinary,
+        [
+          'run',
+          'test',
+          '--precompiled',
+          precompiledPath,
+        ]..addAll(extraTestArgs));
+    var testExitCode = await testProcess.exitCode;
+    if (testExitCode != 0) {
+      // No need to log - should see failed tests in the console.
+      exit(testExitCode);
+    }
   }
 }